refactor(validators): share deduplicateErrors and reuse the isObject helper
Extract the duplicated message+path deduplication into a single
`deduplicateErrors` in @scalar/json-schema-validator, used by both the error
transformer and the path-parameter checks. Replace the local `isRecord` in
validate-path-parameters with the shared `@scalar/helpers` isObject.
