+++ packages/json-schema-validator/src/index.ts
@@ -1,3 +1,4 @@
+export { deduplicateErrors } from '@/deduplicate-errors'
 export { type AjvError, type PrettyError, prettifyAjvErrors } from '@/prettify-ajv-errors'
 export type { CreateValidatorOptions, ErrorObject, ValidateOptions, ValidationResult } from '@/types'
 export { createValidator, validate } from '@/validate'
